patch 8.2.0709: MS-Windows: compiler warning for int vs size_t
Problem:    MS-Windows: compiler warning for int vs size_t.
Solution:   Add type cast. (Mike Williams)

+++ src/channel.c
@@ -1037,7 +1037,7 @@ channel_open(
 	if (waittime == 0)
 	    waittime = 1;
 
-	sd = channel_connect(channel, addr->ai_addr, addr->ai_addrlen,
+	sd = channel_connect(channel, addr->ai_addr, (int)addr->ai_addrlen,
 								   &waittime);
 	if (sd >= 0)
 	    break;
